HTM-MachineLearning-LightWeight-main/requirements.txt
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
google-cloud-speech
google-cloud-texttospeech
fastapi
uvicorn
python-multipart
pillow
python-dotenv